movendo definições para o header da VM
